Allow for more than 2 digits
The initial design only allowed a max of 99 to be entered as I wasn't sure how to increase the size of the field as the content got bigger. Implemented some code from Paul so that it now allows a maximum of 6 digits and the field increases with the more digits that are entered.